Identified the bug of missing Escapes.
'key_buffer_timeout' needs to be set at least larger than twice as much as 'sequence_timeout' for Escape to be captured on Unix-like system in 'main.py'. It is because: suppose Escape is pressed before 'display' in 'main.py', 'safe_io' in 'display' will wait 'sequence_timeout' once, and in the next iteration in the main loop, 'get_key_event' will wait for 'sequence_timeout' again. This might be fixed by also making 'safe_io' parsing the key input, but since it may introduce other issues, it is not yet fixed. It is now documented in the 'Keyboard.__init__' that 'key_buffer_timeout' should be larger than '2*sequence_timeout'.