Fixed a bug in 'PostEffect'; Removed 'spatial_effect' from 'PostEffect'; Fixed a bug in 'Renderer'.
Fixed the bug that leads to incorrect color processing when fxaa is on in 'PostEffect'. It is caused by the buffer data swapping and incorrect usage of the buffer. Removed 'spatial_effect' for clearer logic. The user can certainly still apply this effect by adding a separate pass. This should not be too problematic performance-wise, as these effects themselves tend to be expensive so another sequence iteration should not be the ottleneck.